requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
multidict==6.7.1
pillow==12.1.1
playwright==1.58.0
propcache==0.4.1
pyee==13.0.1
requests==2.33.0
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
